* Modularizacion. * Importar solo funciones necesarias si son pocas en Main.hs.
